consolidate dockefiles--remove base
two dockerfiles isn't ideal, and creates a complex release cycle of
image tagging/building, secondary tagging/building, and causing some
slowdown during the Github action run since a docker image has to be
built (40s).

This consolidates the two docker images so that all of the base pronto
functionality can be used standalone, or it can be used in a github
action.

/runner/pronto can be used for github actions and is the default
entrypoint

(global) pronto is still available and will execute the bundled `pronto`
binstub.